@@ -1,56 +1,54 @@
 {
-  "name": "app",
-  "version": "0.1.0",
+  "name": "tag-my-video",
+  "version": "3.0.0",
+  "description": "Video tagging application for sports analysis",
   "private": true,
+  "packageManager": "pnpm@9.1.0",
+  "engines": {
+    "node": ">=18.0.0",
+    "pnpm": ">=9.0.0"
+  },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.3",
-    "@mui/material": "^5.14.3",
-    "@mui/utils": "^5.14.11",
-    "@mui/x-charts": "^6.0.0-alpha.12",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.7.13",
-    "@types/react": "^18.2.23",
-    "@types/react-dom": "^18.2.8",
-    "@types/recharts": "^1.8.25",
-    "@types/video.js": "^7.3.52",
-    "@types/videojs": "^5.16.2",
-    "concurrently": "^8.2.0",
-    "electron-is-dev": "^2.0.0",
-    "electron-localshortcut": "^3.2.1",
-    "electron-reload": "^2.0.0-alpha.1",
-    "electronmon": "^2.0.2",
-    "eslint": "^8.49.0",
-    "eslint-config-prettier": "^9.0.0",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/icons-material": "^5.16.7",
+    "@mui/material": "^5.16.7",
+    "@mui/utils": "^5.16.6",
+    "@mui/x-charts": "^7.18.0",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^20.16.10",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.0",
+    "@types/recharts": "^1.8.29",
+    "@types/video.js": "^7.3.58",
+    "concurrently": "^9.0.1",
+    "electron-is-dev": "^3.0.1",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^9.1.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
-    "minimatch": "^9.0.3",
+    "minimatch": "^10.0.1",
     "path-browserify": "^1.0.1",
-    "prettier": "^3.0.3",
-    "react": "^18.2.0",
-    "react-datepicker": "^4.16.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.14.2",
+    "prettier": "^3.3.3",
+    "react": "^18.3.1",
+    "react-datepicker": "^7.4.0",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "^5.0.1",
-    "recharts": "^2.8.0",
-    "redux": "^4.2.1",
-    "tsc": "^2.0.4",
-    "typescript": "^4.9.5",
+    "recharts": "^2.12.7",
+    "redux": "^5.0.1",
+    "typescript": "^5.4.5",
     "ulid": "^2.3.0",
-    "video-react": "^0.16.0",
-    "video.js": "^8.3.0",
-    "wait-on": "^7.0.1",
-    "web-audio-api": "^0.2.2",
-    "web-vitals": "^2.1.0"
+    "video.js": "^8.18.1",
+    "wait-on": "^8.0.1",
+    "web-vitals": "^4.2.3"
   },
   "main": "build/electron/src/main.js",
   "homepage": "./",
-  "description": "description",
-  "author": "author",
+  "author": "Kou-ISK",
   "build": {
     "extends": null,
     "files": [
@@ -95,9 +93,11 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@types/react-datepicker": "^4.15.0",
-    "electron": "^25.4.0",
-    "electron-builder": "^24.6.4"
+    "@types/react-datepicker": "^7.0.0",
+    "@typescript-eslint/eslint-plugin": "^7.18.0",
+    "@typescript-eslint/parser": "^7.18.0",
+    "electron": "^31.6.0",
+    "electron-builder": "^25.0.5"
   },
   "repository": "https://github.com/Kou-ISK/tag-my-video.git",
   "license": "MIT"
